Remove Card Kingdom pricelist supplement from CK price refresh
CK prices already come from MTGJSON. The optional live pricelist fetch was
often blocked with 403 from Lambda IPs and only added face-name aliases for
a narrow set of double-faced cards. Drop the supplement path, related fetch
helpers, and CK API product parsing code.
